Rename Blog to BlogList route.
